+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/cpuid.cpp
@@ -23,7 +23,7 @@ CommandCpuidHelp()
                  "cpuids instructions.\n\n");
 
     ShowMessages("syntax : \t!cpuid [Eax (hex)] [pid ProcessId (hex)] [core CoreId (hex)] "
-                 "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] "
                  "[code { Code (hex) }]\n");