+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/crwrite.cpp
@@ -22,7 +22,7 @@ CommandCrwriteHelp()
     ShowMessages("!crwrite : monitors modification of control registers (CR0 / CR4).\n\n");
 
     ShowMessages("syntax : \t!crwrite [Cr (hex)] [mask Mask (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");