+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/ioout.cpp
@@ -23,7 +23,7 @@ CommandIooutHelp()
                  "instructions.\n\n");
 
     ShowMessages("syntax : \t!ioout [Port (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");