+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/msrread.cpp
@@ -22,7 +22,7 @@ CommandMsrreadHelp()
     ShowMessages("!msrread : detects the execution of rdmsr instructions.\n\n");
 
     ShowMessages("syntax : \t!msrread [Msr (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");