+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/msrwrite.cpp
@@ -22,7 +22,7 @@ CommandMsrwriteHelp()
     ShowMessages("!msrwrite : detects the execution of wrmsr instructions.\n\n");
 
     ShowMessages("syntax : \t!msrwrite [Msr (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");