+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/pmc.cpp
@@ -22,9 +22,8 @@ CommandPmcHelp()
     ShowMessages("!pmc : monitors execution of rdpmc instructions.\n\n");
 
     ShowMessages("syntax : \t!pmc [pid ProcessId (hex)] [core CoreId (hex)] "
-                 "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
-                 "[script { Script (string) }] [condition { Condition (hex) }] "
-                 "[code { Code (hex) }]\n");
+                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
+                 "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : !pmc\n");