+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/vmcall.cpp
@@ -22,9 +22,8 @@ CommandVmcallHelp()
     ShowMessages("!vmcall : monitors execution of VMCALL instruction.\n\n");
 
     ShowMessages("syntax : \t!vmcall [pid ProcessId (hex)] [core CoreId (hex)] "
-                 "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
-                 "[script { Script (string) }] [condition { Condition (hex) }] "
-                 "[code { Code (hex) }]\n");
+                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
+                 "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : !vmcall\n");